requirements.txt
# Core dependencies
streamlit
langchain-core>=0.2.0
langchain-text-splitters>=0.2.0
langchain-groq>=0.1.0
langchain-google-genai>=2.0.0
langgraph
tavily-python
langchain-community

# Vector database & embeddings
fastembed

# Document parsing
PyPDF2>=3.0.0
python-docx>=1.1.0

# Utilities
python-dotenv>=1.0.1
typing-extensions>=4.9.0

# arXiv integration
arxiv>=2.1.0

# Telegram & Cloud DB
python-telegram-bot>=21.0
supabase>=2.3.0
